changed file structure
removed schema.py, config.yaml, data.json, outputs

put main, functions into sm folder

put tests into test folder

edit: swapped json_dumps with model_dumps_json to allow for nested schemas
