--- data.json
@@ -1,6 +1,0 @@
-{
-	"street":"Oreilly",
-	"city":"Van Buren",
-	"zip_code":"92071",
-	"country":"Austria"
-}
