--- outputs/output.json
@@ -1,62 +1,0 @@
-[1: {
-        "street": "Crown",
-        "city": "Baldwin",
-        "zip_code": "79034",
-        "country": "Ireland",
-        "username": "delivering_1816"
-}2: {
-        "street": "Laurel",
-        "city": "Muncie",
-        "zip_code": "83849",
-        "country": "Cocos",
-        "username": "holds_1842"
-}3: {
-        "street": "Glendale",
-        "city": "Brooklyn Center",
-        "zip_code": "61609",
-        "country": "Bulgaria",
-        "username": "irrigation_1925"
-}4: {
-        "street": "States",
-        "city": "Greeley",
-        "zip_code": "51387",
-        "country": "Malta",
-        "username": "north_1952"
-}5: {
-        "street": "Hugo",
-        "city": "Parkersburg",
-        "zip_code": "33628",
-        "country": "Libya",
-        "username": "silver_1969"
-}6: {
-        "street": "Ortega",
-        "city": "Stamford",
-        "zip_code": "27964",
-        "country": "Mexico",
-        "username": "transition_1983"
-}7: {
-        "street": "Norfolk",
-        "city": "Helena",
-        "zip_code": "26969",
-        "country": "Luxembourg",
-        "username": "martha_1979"
-}8: {
-        "street": "Bowman",
-        "city": "Pittsburgh",
-        "zip_code": "28963",
-        "country": "Hungary",
-        "username": "sufficiently_2081"
-}9: {
-        "street": "Kohler",
-        "city": "Temple Terrace",
-        "zip_code": "88586",
-        "country": "Svalbard & Jan Mayen",
-        "username": "start_1819"
-}10: {
-        "street": "Rosella",
-        "city": "Helena",
-        "zip_code": "43675",
-        "country": "Iraq",
-        "username": "situation_1887"
-}
-]
