+++ tests/test_synthesiser.py
